refactor: move more functionality from PagesTable into generic components
