+++ www/admin/app/dashboard/pages/PagesTable.tsx
@@ -1,57 +1,33 @@
 "use client";
 
 import * as React from "react";
-import {
-	type ColumnDef,
-	getCoreRowModel,
-	useReactTable,
-	getFilteredRowModel,
-	getSortedRowModel,
-	getPaginationRowModel,
-	flexRender,
-} from "@tanstack/react-table";
-import {
-	Stepper,
-	Table,
-	TableBody,
-	TableCell,
-	TableHead,
-	TableHeader,
-	TableRow,
-} from "@admin/src/components/ui/server";
-import { ChevronRightIcon } from "@heroicons/react/20/solid";
 import type { Page } from "@prisma/client";
-import { cn } from "@admin/src/utils/styling";
-import { ActionIcon, Button, Input } from "@admin/src/components/ui/client";
-import {
-	DocumentPlusIcon,
-	FolderPlusIcon,
-	HomeIcon,
-	MagnifyingGlassIcon,
-} from "@heroicons/react/24/outline";
-import Link from "next/link";
+import { Button } from "@admin/src/components/ui/client";
+import { DocumentPlusIcon, FolderPlusIcon } from "@heroicons/react/24/outline";
 import { trpc } from "@admin/src/utils/trpc";
 import { AddDialog } from "./dialogs";
-import { DataTablePagination } from "@admin/src/components";
+import { DataTable, DataTableBreadcrumbs, DataTablePagination } from "@admin/src/components";
 import type { SearchParams } from "./page";
 import { type TableCookie } from "@admin/src/utils/cookies";
-import { useTableCookie } from "@admin/src/hooks/useTableCookie";
+import { useTable } from "@admin/src/hooks/useTable";
+import { columns } from "./PagesTableColumns";
 
-interface PagesTableProps {
-	columns: ColumnDef<Page>[];
+interface Props {
 	group: Page;
 	data: { pages: Page[]; breadcrumbs: Page[] };
 	pagination: SearchParams;
 	cookie: TableCookie | null;
 }
 
-export function PagesTable(props: PagesTableProps) {
+export function PagesTable(props: Props) {
 	const data = trpc.pages.getGroupContents.useQuery(
 		props.data.breadcrumbs.length > 0 ? { id: props.group.id } : null,
 		{ initialData: props.data },
 	).data;
 
-	const reactTableProps = useTableCookie({
+	const { table, searchElement } = useTable({
+		data: data.pages,
+		columns,
 		cookie: {
 			name: "pages-table",
 			contents: props.cookie,
@@ -62,17 +38,6 @@ export function PagesTable(props: PagesTableProps) {
 		},
 	});
 
-	const table = useReactTable({
-		data: data.pages,
-		columns: props.columns,
-		getCoreRowModel: getCoreRowModel(),
-		getFilteredRowModel: getFilteredRowModel(),
-		getSortedRowModel: getSortedRowModel(),
-		getPaginationRowModel: getPaginationRowModel(),
-		autoResetPageIndex: false,
-		...reactTableProps,
-	});
-
 	const [openAddPage, setOpenAddPage] = React.useState(false);
 	const [openAddGroup, setOpenAddGroup] = React.useState(false);
 
@@ -80,13 +45,7 @@ export function PagesTable(props: PagesTableProps) {
 		<>
 			<div className="flex flex-col gap-4" data-testid="pages-table">
 				<div className="flex justify-between gap-2">
-					<Input
-						type="search"
-						className="mr-auto max-w-xs"
-						value={(table.getColumn("name")?.getFilterValue() as string) ?? ""}
-						onChange={(e) => table.getColumn("name")?.setFilterValue(e.target.value)}
-						icon={<MagnifyingGlassIcon className="w-4" />}
-					/>
+					{searchElement}
 
 					<div className="flex gap-2">
 						<Button
@@ -107,99 +66,8 @@ export function PagesTable(props: PagesTableProps) {
 					</div>
 				</div>
 
-				{data.breadcrumbs.length > 0 && (
-					<Stepper
-						className="-mb-2"
-						steps={[
-							<Link key={0} href={"/dashboard/pages"}>
-								<ActionIcon className="-mr-2">
-									<HomeIcon className="w-5 text-foreground" />
-								</ActionIcon>
-							</Link>,
-							...data.breadcrumbs.map((breadcrumb, i) => (
-								<Button
-									key={i + 1}
-									variant={"link"}
-									className={cn(
-										"whitespace-nowrap font-normal",
-										i < data.breadcrumbs.length - 1 && "text-muted-foreground",
-									)}
-									asChild
-								>
-									<Link key={i} href={`/dashboard/pages/${breadcrumb.id}`}>
-										{breadcrumb.name}
-									</Link>
-								</Button>
-							)),
-						]}
-						currentStep={data.breadcrumbs.length}
-						separator={<ChevronRightIcon className="w-4" />}
-						data-testid="breadcrumbs"
-					/>
-				)}
-
-				<div className="rounded-md border">
-					<Table>
-						<TableHeader>
-							{table.getHeaderGroups().map((headerGroup) => (
-								<TableRow
-									key={headerGroup.id}
-									className="whitespace-nowrap hover:bg-inherit"
-								>
-									{headerGroup.headers.map((header) => (
-										<TableHead
-											key={header.id}
-											style={{ width: header.getSize().toString() + "px" }}
-										>
-											{header.isPlaceholder
-												? null
-												: flexRender(
-														header.column.columnDef.header,
-														header.getContext(),
-												  )}
-										</TableHead>
-									))}
-								</TableRow>
-							))}
-						</TableHeader>
-
-						<TableBody>
-							{table.getRowModel().rows?.length ? (
-								table.getRowModel().rows.map((row) => (
-									<TableRow
-										key={row.id}
-										data-state={row.getIsSelected() && "selected"}
-									>
-										{row.getVisibleCells().map((cell, i, cells) => (
-											<TableCell
-												key={cell.id}
-												className={cn(
-													"whitespace-nowrap",
-													i > 0 && "text-muted-foreground",
-													i === cells.length - 1 && "py-0",
-												)}
-											>
-												{flexRender(
-													cell.column.columnDef.cell,
-													cell.getContext(),
-												)}
-											</TableCell>
-										))}
-									</TableRow>
-								))
-							) : (
-								<TableRow>
-									<TableCell
-										colSpan={props.columns.length}
-										className="h-24 text-center"
-									>
-										No results.
-									</TableCell>
-								</TableRow>
-							)}
-						</TableBody>
-					</Table>
-				</div>
+				<DataTableBreadcrumbs breadcrumbs={data.breadcrumbs} />
+				<DataTable table={table} columns={columns} />
 
 				<DataTablePagination table={table} />
 			</div>
