--- www/admin/src/components/DataTablePagination.tsx
@@ -1,88 +1,0 @@
-import type { Table } from "@tanstack/react-table";
-import {
-	ChevronLeftIcon,
-	ChevronRightIcon,
-	ChevronDoubleLeftIcon,
-	ChevronDoubleRightIcon,
-} from "@heroicons/react/20/solid";
-import {
-	Select,
-	SelectContent,
-	SelectItem,
-	SelectTrigger,
-	SelectValue,
-	Button,
-} from "@admin/src/components/ui/client";
-
-interface DataTablePaginationProps<TData> {
-	table: Table<TData>;
-}
-export function DataTablePagination<TData>({ table }: DataTablePaginationProps<TData>) {
-	return (
-		<div className="ml-auto flex flex-wrap items-center justify-between">
-			<div className="flex items-center gap-6">
-				<div className="flex items-center space-x-2">
-					<p className="text-sm font-medium">Rows per page</p>
-					<Select
-						value={table.getState().pagination.pageSize.toString()}
-						onValueChange={(value) => {
-							table.setPageSize(Number(value));
-						}}
-					>
-						<SelectTrigger className="h-8 w-[70px]">
-							<SelectValue placeholder={table.getState().pagination.pageSize} />
-						</SelectTrigger>
-						<SelectContent side="top">
-							{[10, 20, 30, 40, 50].map((pageSize) => (
-								<SelectItem key={pageSize} value={`${pageSize}`}>
-									{pageSize}
-								</SelectItem>
-							))}
-						</SelectContent>
-					</Select>
-				</div>
-				<div className="flex w-[100px] items-center justify-center text-sm font-medium">
-					Page {table.getState().pagination.pageIndex + 1} of {table.getPageCount()}
-				</div>
-				<div className="flex items-center space-x-2">
-					<Button
-						variant="outline"
-						className="hidden h-8 w-8 p-0 lg:flex"
-						onClick={() => table.setPageIndex(0)}
-						disabled={!table.getCanPreviousPage()}
-					>
-						<span className="sr-only">Go to first page</span>
-						<ChevronDoubleLeftIcon className="h-4 w-4" />
-					</Button>
-					<Button
-						variant="outline"
-						className="h-8 w-8 p-0"
-						onClick={() => table.previousPage()}
-						disabled={!table.getCanPreviousPage()}
-					>
-						<span className="sr-only">Go to previous page</span>
-						<ChevronLeftIcon className="h-4 w-4" />
-					</Button>
-					<Button
-						variant="outline"
-						className="h-8 w-8 p-0"
-						onClick={() => table.nextPage()}
-						disabled={!table.getCanNextPage()}
-					>
-						<span className="sr-only">Go to next page</span>
-						<ChevronRightIcon className="h-4 w-4" />
-					</Button>
-					<Button
-						variant="outline"
-						className="hidden h-8 w-8 p-0 lg:flex"
-						onClick={() => table.setPageIndex(table.getPageCount() - 1)}
-						disabled={!table.getCanNextPage()}
-					>
-						<span className="sr-only">Go to last page</span>
-						<ChevronDoubleRightIcon className="h-4 w-4" />
-					</Button>
-				</div>
-			</div>
-		</div>
-	);
-}
