+++ www/admin/src/components/index.ts
@@ -1,5 +1,5 @@
 export { Body } from "./Body";
 export { Combobox, type ComboboxData } from "./Combobox";
-export { DataTablePagination } from "./DataTablePagination";
+export { DataTable, DataTablePagination, DataTableBreadcrumbs } from "./DataTable";
 export { InfoTooltip } from "./InfoTooltip";
 export { PageTitle } from "./PageTitle";
